refactor: 단어 추가 버튼을 TopAppBar로 옮기고, TopAppBar를 여러 개의 composable로 나눔
