+++ app/src/main/java/hsk/practice/myvoca/ui/MyVocaApp.kt
@@ -1,17 +1,13 @@
 package hsk.practice.myvoca.ui
 
-import android.content.Intent
 import androidx.compose.animation.ExperimentalAnimationApi
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.*
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.outlined.Add
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Scaffold
+import androidx.compose.material.primarySurface
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.SideEffect
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
@@ -21,7 +17,6 @@ import androidx.navigation.compose.rememberNavController
 import com.google.accompanist.systemuicontroller.rememberSystemUiController
 import hsk.practice.myvoca.ui.components.MyVocaBottomAppBar
 import hsk.practice.myvoca.ui.components.MyVocaTopAppBar
-import hsk.practice.myvoca.ui.screens.addword.AddWordActivity
 import hsk.practice.myvoca.ui.screens.allword.AllWordScreen
 import hsk.practice.myvoca.ui.screens.allword.AllWordViewModel
 import hsk.practice.myvoca.ui.screens.home.HomeScreen
@@ -50,11 +45,9 @@ fun MyVocaApp() {
             )
         }
 
-        val context = LocalContext.current
-
         Scaffold(
             topBar = {
-                MyVocaTopAppBar()
+                MyVocaTopAppBar(currentScreen = currentScreen)
             },
             bottomBar = {
                 MyVocaBottomAppBar(
@@ -70,22 +63,6 @@ fun MyVocaApp() {
                     },
                     currentScreen = currentScreen
                 )
-            },
-            floatingActionButton = {
-                IconButton(
-                    onClick = {
-                        context.startActivity(Intent(context, AddWordActivity::class.java))
-                    },
-                    modifier = Modifier.background(
-                        color = MaterialTheme.colors.secondary,
-                        shape = CircleShape
-                    )
-                ) {
-                    Icon(
-                        imageVector = Icons.Outlined.Add,
-                        contentDescription = "새로운 단어를 추가할 수 있습니다.",
-                    )
-                }
             },
         ) { innerPadding ->
             MyVocaNavGraph(
