+++ app/src/main/java/hsk/practice/myvoca/ui/components/InsetAwareTopAppBar.kt
@@ -1,0 +1,46 @@
+package hsk.practice.myvoca.ui.components
+
+import androidx.compose.foundation.layout.RowScope
+import androidx.compose.material.*
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+import com.google.accompanist.insets.navigationBarsPadding
+import com.google.accompanist.insets.statusBarsPadding
+
+/**
+ * From Jetnews, Google's official sample app.
+ *
+ * A wrapper around [TopAppBar] which uses [Modifier.statusBarsPadding] to shift the app bar's
+ * contents down, but still draws the background behind the status bar too.
+ */
+@Composable
+fun InsetAwareTopAppBar(
+    title: @Composable () -> Unit,
+    modifier: Modifier = Modifier,
+    navigationIcon: @Composable (() -> Unit)? = null,
+    actions: @Composable RowScope.() -> Unit = {},
+    backgroundColor: Color = MaterialTheme.colors.primarySurface,
+    contentColor: Color = contentColorFor(backgroundColor),
+    elevation: Dp = 4.dp
+) {
+    Surface(
+        color = backgroundColor,
+        elevation = elevation,
+        modifier = modifier
+    ) {
+        TopAppBar(
+            title = title,
+            navigationIcon = navigationIcon,
+            actions = actions,
+            backgroundColor = Color.Transparent,
+            contentColor = contentColor,
+            elevation = 0.dp,
+            modifier = Modifier
+                .statusBarsPadding()
+                .navigationBarsPadding(bottom = false),
+        )
+    }
+}
